+++ NEWS.md
@@ -1,6 +1,8 @@
-# amlrDatabases 0.6.2.9000
+# amlrDatabases 0.7.0.9000
 
-* Updated `mod_output` to more consistently display plots, tables, and associated widgets and messages
+* Removed `parent` argument from `mod_output_server`. The needed information (i.e., the string of the parent namespace ID) is now determined using `session`
+
+* Updated `mod_output` to more consistently display plots, tables, and associated widgets and messages, through the use of `reactiveVal` and `renderUI`/`uiOutput` functions
 
 
 # amlrDatabases 0.6.1